make as window
